+++ src/main/kotlin/Main.kt
@@ -3,6 +3,6 @@ import javafx.application.Application
 
 fun main(args : Array<String>) {
     println("JavaFX")
-    Application.launch(GUI().javaClass, *args)
+    Application.launch(TAB().javaClass, *args)
     println("End")
 }