+++ ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased changes
 
+### Add
+* TabView
+* TableView with Checkbox for grouping files
+
 ## 0.1.0.0  -- 2018.12.27
 
 Implement default threaded code